+++ forum/Controllers/PostController.cs
@@ -19,9 +19,42 @@ namespace forum.Controllers
             _context = context;
         }
 
+        private async Task<User?> GetUserAsync()
+        {
+            string? token = HttpContext.Session.GetString("token");
+            if (token == null)
+            {
+                return null;
+            }
+            var user = await _context.User.FirstOrDefaultAsync(u => u.token == token);
+            if (user == null)
+            {
+                return null;
+            }
+            return user;
+        }
+
+
         // GET: Post
         public async Task<IActionResult> Index()
         {
+            var userSession = await GetUserAsync();
+            if (userSession == null)
+            {
+                return RedirectToAction("SignIn", "Auth");
+            }
+
+            var posts = await _context.Post.ToListAsync();
+            var idToName = new Dictionary<int, string>();
+            foreach (var post in posts)
+            {
+                var user = await _context.User.FindAsync(post.userId);
+                if (user != null)
+                    idToName[post.userId] = user.name + " " + user.lastName;
+            }
+
+            ViewData["idToName"] = idToName;
+
             return _context.Post != null ?
                         View(await _context.Post.ToListAsync()) :
                         Problem("Entity set 'ForumDbContext.Post'  is null.");
@@ -30,6 +63,12 @@ namespace forum.Controllers
         // GET: Post/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            var userSession = await GetUserAsync();
+            if (userSession == null)
+            {
+                return RedirectToAction("SignIn", "Auth");
+            }
+
             if (id == null || _context.Post == null)
             {
                 return NotFound();
@@ -46,8 +85,14 @@ namespace forum.Controllers
         }
 
         // GET: Post/Create
-        public IActionResult Create()
+        public async Task<IActionResult> Create()
         {
+            var userSession = await GetUserAsync();
+            if (userSession == null)
+            {
+                return RedirectToAction("SignIn", "Auth");
+            }
+
             return View();
         }
 
@@ -56,32 +101,47 @@ namespace forum.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("id,userId,title,content,createdAt")] Post post)
+        public async Task<IActionResult> Create(IFormCollection postForm)
         {
+            var userSession = await GetUserAsync();
+            if (userSession == null)
+            {
+                return RedirectToAction("SignIn", "Auth");
+            }
+
+            var tag = _context.Tag.FirstOrDefault(t => t.name == postForm["tag"].ToString());
+            if (tag == null)
+            {
+                tag = new Tag
+                {
+                    name = postForm["tag"].ToString(),
+                };
+            }
+
+            var post = new Post
+            {
+                userId = userSession.id,
+                title = postForm["title"].ToString(),
+                content = postForm["content"].ToString(),
+                tagId = tag.id,
+            };
+
             if (ModelState.IsValid)
             {
-                string? token = HttpContext.Session.GetString("token");
-                if (token == null)
-                {
-                    return RedirectToAction("SignIn", "Auth");
-                }
-                var user = _context.User.FirstOrDefault(u => u.token == token);
-                if (user == null)
-                {
-                    return Problem("User does not exist.");
-                }
 
-                post.userId = user.id;
                 _context.Add(post);
                 await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("Index", "Home");
             }
+            
             return View(post);
         }
 
         // GET: Post/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            var userSession = await GetUserAsync();
+
             if (id == null || _context.Post == null)
             {
                 return NotFound();
@@ -92,6 +152,12 @@ namespace forum.Controllers
             {
                 return NotFound();
             }
+
+            if (userSession == null || userSession.role != "admin" && userSession.id != post.userId)
+            {
+                return RedirectToAction("SignIn", "Auth");
+            }
+
             return View(post);
         }
 
@@ -105,6 +171,12 @@ namespace forum.Controllers
             if (id != post.id)
             {
                 return NotFound();
+            }
+
+            var userSession = await GetUserAsync();
+            if (userSession == null || userSession.role != "admin" && userSession.id != post.userId)
+            {
+                return RedirectToAction("SignIn", "Auth");
             }
 
             if (ModelState.IsValid)
@@ -138,11 +210,18 @@ namespace forum.Controllers
                 return NotFound();
             }
 
+            var userSession = await GetUserAsync();
+
             var post = await _context.Post
                 .FirstOrDefaultAsync(m => m.id == id);
             if (post == null)
             {
                 return NotFound();
+            }
+
+            if (userSession == null || userSession.role != "admin" && userSession.id != post.userId)
+            {
+                return RedirectToAction("SignIn", "Auth");
             }
 
             return View(post);
@@ -157,11 +236,20 @@ namespace forum.Controllers
             {
                 return Problem("Entity set 'ForumDbContext.Post'  is null.");
             }
+
+            var userSession = await GetUserAsync();
             var post = await _context.Post.FindAsync(id);
+
             if (post != null)
             {
+                if (userSession == null || userSession.role != "admin" && userSession.id != post.userId)
+                {
+                    return RedirectToAction("SignIn", "Auth");
+                }
+
                 _context.Post.Remove(post);
             }
+
 
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
@@ -174,6 +262,12 @@ namespace forum.Controllers
 
         public async Task<IActionResult> GiveLike(int id)
         {
+            var user = await GetUserAsync();
+            if (user == null)
+            {
+                return RedirectToAction("SignIn", "Auth");
+            }
+
             var post = _context.Post.Where(p => p.id == id).ToList()[0];
             post.likes += 1;
             _context.Update(post);
@@ -183,6 +277,12 @@ namespace forum.Controllers
 
         public async Task<IActionResult> GiveDislike(int id)
         {
+            var user = await GetUserAsync();
+            if (user == null)
+            {
+                return RedirectToAction("SignIn", "Auth");
+            }
+
             var post = _context.Post.Where(p => p.id == id).ToList()[0];
             post.dislikes += 1;
             _context.Update(post);
